pyproject.toml
[project]
name = "azure-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Aidar Darmesh",email = "darmesh.aidar@gmail.com"}
]
readme = "README.md"
requires-python = "3.12"
dependencies = [
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "azure-functions (>=1.23.0,<2.0.0)",
    "azure-mgmt-signalr (>=1.2.0,<2.0.0)",
    "azure-messaging-webpubsubservice (>=1.3.0,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
azure-ai-projects = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
